requirements.txt
Package                   Version      Editable project location
------------------------- ------------ -------------------------
absl-py                   2.3.1
aiofiles                  23.2.1
altair                    5.5.0
annotated-doc             0.0.4
anyio                     4.11.0
attrs                     25.4.0
blurhash                  1.1.5
boost                     0.1
captum                    0.8.0
certifi                   2025.11.12
charset-normalizer        3.4.4
click                     8.1.8
cmake                     4.1.2
contourpy                 1.3.0
cycler                    0.12.1
decorator                 5.2.1
dlib                      20.0.0
einops                    0.8.1
exceptiongroup            1.3.0
fastapi                   0.121.2
ffmpy                     1.0.0
filelock                  3.19.1
fonttools                 4.60.1
fsspec                    2025.9.0
gradio                    3.50.2
gradio_client             0.6.1
greenlet                  3.2.4
grpcio                    1.76.0
h11                       0.16.0
hf-xet                    1.2.0
httpcore                  1.0.9
httpx                     0.28.1
huggingface_hub           1.1.4
idna                      3.11
importlib_metadata        8.7.0
importlib_resources       6.5.2
Jinja2                    3.1.6
joblib                    1.5.2
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
kiwisolver                1.4.7
Markdown                  3.9
MarkupSafe                2.1.5
Mastodon.py               2.1.4
matplotlib                3.9.4
mpmath                    1.3.0
narwhals                  2.11.0
networkx                  3.2.1
neurovisus                1.0.0        /mnt/e/NeuroVisus
numpy                     2.0.2
nvidia-cublas-cu11        11.11.3.6
nvidia-cuda-cupti-cu11    11.8.87
nvidia-cuda-nvrtc-cu11    11.8.89
nvidia-cuda-runtime-cu11  11.8.89
nvidia-cudnn-cu11         9.1.0.70
nvidia-cufft-cu11         10.9.0.58
nvidia-curand-cu11        10.3.0.86
nvidia-cusolver-cu11      11.4.1.48
nvidia-cusparse-cu11      11.7.5.86
nvidia-nccl-cu11          2.21.5
nvidia-nvtx-cu11          11.8.86
opencv-python             4.12.0.88
orjson                    3.11.4
packaging                 25.0
pandas                    2.3.3
pillow                    10.4.0
pip                       25.3
protobuf                  6.33.1
pydantic                  1.10.24
pydub                     0.25.1
pyparsing                 3.2.5
python-dateutil           2.9.0.post0
python-magic              0.4.27
python-multipart          0.0.20
pytz                      2025.2
PyYAML                    6.0.3
referencing               0.36.2
requests                  2.32.5
rpds-py                   0.27.1
scikit-learn              1.6.1
scipy                     1.13.1
seaborn                   0.13.2
semantic-version          2.10.0
setuptools                80.9.0
shellingham               1.5.4
six                       1.17.0
sniffio                   1.3.1
SQLAlchemy                2.0.44
starlette                 0.49.3
sympy                     1.14.0
tensorboard               2.20.0
tensorboard-data-server   0.7.2
threadpoolctl             3.6.0
torch                     2.7.1+cu118
torchaudio                2.7.1+cu118
torchvision               0.22.1+cu118
tqdm                      4.67.1
triton                    3.3.1
typer-slim                0.20.0
typing_extensions         4.15.0
tzdata                    2025.2
urllib3                   2.5.0
uvicorn                   0.38.0
websockets                11.0.3
Werkzeug                  3.1.3
wheel                     0.45.1
zipp                      3.23.0